fix(mtrbasis): constant splines use half-open intervals
Using a closed interval on the right and left means that at points of
discontinuity, we double-count it twice.
This fix ensures that we only count such points 1.
I opted to use half-open intervals that are closed on the left and open
on the right.
However, the final interval will effectively be closed on the right
because I added 0.1 to the last knot.
This seems a bit hacky but should work.