fix(utils): don't copy folder if destination already exists
If I want to reproduce/overwrite the images in an existing results/
directory, I won't be able to unless I delete results/tex/*.
This is because the tex/* directory won't be copied to an existing
results/tex/* directory.

This commit first checks to see if results/tex/* directory exists.
If it does, then the tex/ directory won't be copied.
This is consistent with other directories that we copy.